+++ domain/src/main/java/com/susu/domain/repository/LoginRepository.kt
@@ -1,0 +1,9 @@
+package com.susu.domain.repository
+
+import com.susu.core.model.Token
+
+interface LoginRepository {
+    suspend fun login(oauthAccessToken: String): Token
+    suspend fun logout()
+    suspend fun withdraw()
+}
